Refactor test suite by pulling Cardano specifics
