+++ ouroboros-network/sim-tests/Main.hs
@@ -14,12 +14,13 @@ import Test.Ouroboros.Network.MockNode qualified (tests)
 import Test.Ouroboros.Network.NodeToClient.Version qualified (tests)
 import Test.Ouroboros.Network.NodeToNode.Version qualified (tests)
 import Test.Ouroboros.Network.PeerSelection qualified (tests)
+import Test.Ouroboros.Network.PeerSelection.Cardano.LocalRootPeers qualified
+import Test.Ouroboros.Network.PeerSelection.Cardano.MockEnvironment qualified
+import Test.Ouroboros.Network.PeerSelection.Cardano.PublicRootPeers qualified
 import Test.Ouroboros.Network.PeerSelection.Json qualified (tests)
 import Test.Ouroboros.Network.PeerSelection.KnownPeers qualified
 import Test.Ouroboros.Network.PeerSelection.LocalRootPeers qualified
-import Test.Ouroboros.Network.PeerSelection.MockEnvironment qualified
 import Test.Ouroboros.Network.PeerSelection.PeerMetric qualified
-import Test.Ouroboros.Network.PeerSelection.PublicRootPeers qualified
 import Test.Ouroboros.Network.PeerSelection.RootPeersDNS qualified
 import Test.Ouroboros.Network.PeerState qualified (tests)
 import Test.Ouroboros.Network.Version qualified (tests)
@@ -35,7 +36,6 @@ tests =
 
     -- network logic
   , Test.Ouroboros.Network.Diffusion.Policies.tests
-  , Test.Ouroboros.Network.Diffusion.Testnet.Cardano.tests
   , Test.Ouroboros.Network.LedgerPeers.tests
   , Test.Ouroboros.Network.MiniProtocols.BlockFetch.tests
   , Test.Ouroboros.Network.MiniProtocols.KeepAlive.tests
@@ -45,13 +45,18 @@ tests =
   , Test.Ouroboros.Network.PeerSelection.Json.tests
   , Test.Ouroboros.Network.PeerSelection.KnownPeers.tests
   , Test.Ouroboros.Network.PeerSelection.LocalRootPeers.tests
-  , Test.Ouroboros.Network.PeerSelection.MockEnvironment.tests
   , Test.Ouroboros.Network.PeerSelection.PeerMetric.tests
-  , Test.Ouroboros.Network.PeerSelection.PublicRootPeers.tests
   , Test.Ouroboros.Network.PeerSelection.RootPeersDNS.tests
   , Test.Ouroboros.Network.PeerSelection.tests
   , Test.Ouroboros.Network.PeerState.tests
   , Test.Ouroboros.Network.Version.tests
+
+    -- cardano specific logic
+  , Test.Ouroboros.Network.Diffusion.Testnet.Cardano.tests
+  , Test.Ouroboros.Network.PeerSelection.Cardano.LocalRootPeers.tests
+  , Test.Ouroboros.Network.PeerSelection.Cardano.MockEnvironment.tests
+  , Test.Ouroboros.Network.PeerSelection.Cardano.PublicRootPeers.tests
+
 
     -- pseudo system-level
   , Test.Ouroboros.Network.MockNode.tests
